Settle on new standard for datatypes throughout pipeline. New read functions leave images as uints. Resize methods return images to uints. Preprocessing functions cast expected uint inputs into float32.

+++ datasets/io/read_image/read_dicom.py
@@ -8,7 +8,7 @@ def read_dicom_uint16(path):
                                           dtype=tf.uint16, scale='auto',
                                           on_error='strict')[0]
     image = tf.image.grayscale_to_rgb(tf.cast(image, tf.int32))
-    image = tf.cast(image, tf.float32)
+    image = tf.cast(image, tf.uint16)
     
     return image
 
@@ -17,7 +17,7 @@ def read_dicom_uint8(path):
     image = tfio.image.decode_dicom_image(image, color_dim=True,
                                           dtype=tf.uint8, scale='auto',
                                           on_error='lossy')[0]
-    image = tf.image.grayscale_to_rgb(tf.cast(image, tf.int32))
-    image = tf.cast(image, tf.float32)
+    image = tf.image.grayscale_to_rgb(image)
+    image = tf.cast(image, tf.uint8)
     
     return image
